requirements.txt
pydantic==2.5.3
pydantic-settings==2.1.0
openai~=1.6.1
pyodbc == 5.0.1
bs4
dash-bootstrap-components
dash~=2.17.0
dash-extensions  
gunicorn~=22.0.0
fastapi==0.108.0
sqlmodel == 0.0.14
langchain-core == 0.1.42
langchain-openai < 2.0.0
langchain-anthropic == 0.1.8
langchain == 0.1.16
openpyxl
qdrant_client
nbformat>=4.2.0
azure-appconfiguration==1.5.0
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0